FileSys: abstract SD save data archive source
